api/pongo/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.3.0
asn1crypto==1.5.1
attrs==25.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==39.0.2
databases==0.9.0
dataclass-factory==2.16
fastapi==0.114.0
frozenlist==1.5.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.1
iap_local_receipt==0.2.0
idna==3.8
multidict==6.1.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
propcache==0.3.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyJWT==2.10.1
pyOpenSSL==23.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.14.1
shazamio==0.8.0
shazamio_core==1.1.2
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
spotipy==2.24.0
SQLAlchemy==2.0.36
starlette==0.38.4
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
yarl==1.18.3
yt-dlp==2025.1.15
